requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.6.1
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.0
